Remove the newer SqlQueryScheduler
Remove the "new" SqlQueryScheduler, which was created to enable stage
retries, but was never rolled out and is no longer important because of
Presto-on-Spark. Rename LegacySqlQueryScheduler to SqlQueryScheduler.
