+++ presto-main/src/main/java/com/facebook/presto/SystemSessionProperties.java
@@ -98,7 +98,6 @@ public final class SystemSessionProperties
     public static final String GROUPED_EXECUTION = "grouped_execution";
     public static final String RECOVERABLE_GROUPED_EXECUTION = "recoverable_grouped_execution";
     public static final String MAX_FAILED_TASK_PERCENTAGE = "max_failed_task_percentage";
-    public static final String MAX_STAGE_RETRIES = "max_stage_retries";
     public static final String PREFER_STREAMING_OPERATORS = "prefer_streaming_operators";
     public static final String TASK_WRITER_COUNT = "task_writer_count";
     public static final String TASK_PARTITIONED_WRITER_COUNT = "task_partitioned_writer_count";
@@ -208,7 +207,6 @@ public final class SystemSessionProperties
     public static final String LIST_BUILT_IN_FUNCTIONS_ONLY = "list_built_in_functions_only";
     public static final String PARTITIONING_PRECISION_STRATEGY = "partitioning_precision_strategy";
     public static final String EXPERIMENTAL_FUNCTIONS_ENABLED = "experimental_functions_enabled";
-    public static final String USE_LEGACY_SCHEDULER = "use_legacy_scheduler";
     public static final String OPTIMIZE_COMMON_SUB_EXPRESSIONS = "optimize_common_sub_expressions";
     public static final String PREFER_DISTRIBUTED_UNION = "prefer_distributed_union";
     public static final String WARNING_HANDLING = "warning_handling";
@@ -467,11 +465,6 @@ public final class SystemSessionProperties
                         RECOVERABLE_GROUPED_EXECUTION,
                         "Experimental: Use recoverable grouped execution when possible",
                         featuresConfig.isRecoverableGroupedExecutionEnabled(),
-                        false),
-                integerProperty(
-                        MAX_STAGE_RETRIES,
-                        "Maximum number of times that stages can be retried",
-                        featuresConfig.getMaxStageRetries(),
                         false),
                 booleanProperty(
                         PREFER_STREAMING_OPERATORS,
@@ -1179,11 +1172,6 @@ public final class SystemSessionProperties
                         EXPERIMENTAL_FUNCTIONS_ENABLED,
                         "Enable listing of functions marked as experimental",
                         featuresConfig.isExperimentalFunctionsEnabled(),
-                        false),
-                booleanProperty(
-                        USE_LEGACY_SCHEDULER,
-                        "Use version of scheduler before refactorings for section retries",
-                        featuresConfig.isUseLegacyScheduler(),
                         false),
                 booleanProperty(
                         OPTIMIZE_COMMON_SUB_EXPRESSIONS,
@@ -2038,11 +2026,6 @@ public final class SystemSessionProperties
         return session.getSystemProperty(MAX_FAILED_TASK_PERCENTAGE, Double.class);
     }
 
-    public static int getMaxStageRetries(Session session)
-    {
-        return session.getSystemProperty(MAX_STAGE_RETRIES, Integer.class);
-    }
-
     public static boolean preferStreamingOperators(Session session)
     {
         return session.getSystemProperty(PREFER_STREAMING_OPERATORS, Boolean.class);
@@ -2672,11 +2655,6 @@ public final class SystemSessionProperties
     public static boolean isExperimentalFunctionsEnabled(Session session)
     {
         return session.getSystemProperty(EXPERIMENTAL_FUNCTIONS_ENABLED, Boolean.class);
-    }
-
-    public static boolean isUseLegacyScheduler(Session session)
-    {
-        return session.getSystemProperty(USE_LEGACY_SCHEDULER, Boolean.class);
     }
 
     public static boolean isOptimizeCommonSubExpressions(Session session)
