+++ presto-main/src/main/java/com/facebook/presto/execution/scheduler/SqlQueryScheduler.java
@@ -19,7 +19,6 @@ import com.facebook.airlift.stats.TimeStat;
 import com.facebook.presto.Session;
 import com.facebook.presto.cost.StatsAndCosts;
 import com.facebook.presto.execution.BasicStageExecutionStats;
-import com.facebook.presto.execution.ExecutionFailureInfo;
 import com.facebook.presto.execution.LocationFactory;
 import com.facebook.presto.execution.PartialResultQueryManager;
 import com.facebook.presto.execution.QueryState;
@@ -35,7 +34,6 @@ import com.facebook.presto.execution.TaskId;
 import com.facebook.presto.execution.buffer.OutputBuffers;
 import com.facebook.presto.execution.buffer.OutputBuffers.OutputBufferId;
 import com.facebook.presto.metadata.FunctionAndTypeManager;
-import com.facebook.presto.metadata.InternalNodeManager;
 import com.facebook.presto.metadata.Metadata;
 import com.facebook.presto.spi.PrestoException;
 import com.facebook.presto.spi.VariableAllocator;
@@ -50,18 +48,15 @@ import com.facebook.presto.sql.planner.TypeProvider;
 import com.facebook.presto.sql.planner.optimizations.PlanOptimizer;
 import com.facebook.presto.sql.planner.plan.PlanFragmentId;
 import com.facebook.presto.sql.planner.sanity.PlanChecker;
-import com.google.common.base.VerifyException;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableListMultimap;
 import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.ListMultimap;
+import com.google.common.collect.Iterables;
 import com.google.common.util.concurrent.ListenableFuture;
 import io.airlift.units.DataSize;
 import io.airlift.units.Duration;
 
 import java.net.URI;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -71,29 +66,25 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import static com.facebook.airlift.concurrent.MoreFutures.tryGetFutureValue;
 import static com.facebook.airlift.concurrent.MoreFutures.whenAnyComplete;
 import static com.facebook.airlift.http.client.HttpUriBuilder.uriBuilderFrom;
 import static com.facebook.presto.SystemSessionProperties.getMaxConcurrentMaterializations;
-import static com.facebook.presto.SystemSessionProperties.getMaxStageRetries;
 import static com.facebook.presto.SystemSessionProperties.getPartialResultsCompletionRatioThreshold;
 import static com.facebook.presto.SystemSessionProperties.getPartialResultsMaxExecutionTimeMultiplier;
 import static com.facebook.presto.SystemSessionProperties.isPartialResultsEnabled;
 import static com.facebook.presto.SystemSessionProperties.isRuntimeOptimizerEnabled;
 import static com.facebook.presto.execution.BasicStageExecutionStats.aggregateBasicStageStats;
-import static com.facebook.presto.execution.SqlStageExecution.RECOVERABLE_ERROR_CODES;
-import static com.facebook.presto.execution.StageExecutionInfo.unscheduledExecutionInfo;
+import static com.facebook.presto.execution.StageExecutionState.ABORTED;
 import static com.facebook.presto.execution.StageExecutionState.CANCELED;
 import static com.facebook.presto.execution.StageExecutionState.FAILED;
 import static com.facebook.presto.execution.StageExecutionState.FINISHED;
+import static com.facebook.presto.execution.StageExecutionState.PLANNED;
 import static com.facebook.presto.execution.StageExecutionState.RUNNING;
 import static com.facebook.presto.execution.StageExecutionState.SCHEDULED;
 import static com.facebook.presto.execution.buffer.OutputBuffers.BROADCAST_PARTITION_ID;
@@ -101,14 +92,12 @@ import static com.facebook.presto.execution.buffer.OutputBuffers.createDiscardin
 import static com.facebook.presto.execution.buffer.OutputBuffers.createInitialEmptyOutputBuffers;
 import static com.facebook.presto.execution.scheduler.StreamingPlanSection.extractStreamingSections;
 import static com.facebook.presto.spi.StandardErrorCode.GENERIC_INTERNAL_ERROR;
-import static com.facebook.presto.sql.planner.PlanFragmenterUtils.isRootFragment;
+import static com.facebook.presto.sql.planner.PlanFragmenterUtils.ROOT_FRAGMENT_ID;
 import static com.facebook.presto.sql.planner.SchedulingOrderVisitor.scheduleOrder;
 import static com.facebook.presto.sql.planner.planPrinter.PlanPrinter.jsonFragmentPlan;
-import static com.google.common.base.Preconditions.checkState;
-import static com.google.common.base.Verify.verify;
+import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.collect.ImmutableList.toImmutableList;
 import static com.google.common.collect.ImmutableMap.toImmutableMap;
-import static com.google.common.collect.Iterables.getLast;
 import static com.google.common.collect.Iterables.getOnlyElement;
 import static com.google.common.collect.Streams.stream;
 import static com.google.common.graph.Traverser.forTree;
@@ -116,6 +105,7 @@ import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.concurrent.TimeUnit.SECONDS;
+import static java.util.function.Function.identity;
 
 public class SqlQueryScheduler
         implements SqlQuerySchedulerInterface
@@ -124,54 +114,52 @@ public class SqlQueryScheduler
 
     private final LocationFactory locationFactory;
     private final ExecutionPolicy executionPolicy;
-    private final ExecutorService executor;
-    private final SplitSchedulerStats schedulerStats;
-    private final SectionExecutionFactory sectionExecutionFactory;
-    private final RemoteTaskFactory remoteTaskFactory;
-    private final SplitSourceFactory splitSourceFactory;
-    private final InternalNodeManager nodeManager;
 
-    private final Session session;
+    private final SplitSchedulerStats schedulerStats;
+
     private final QueryStateMachine queryStateMachine;
     private final AtomicReference<SubPlan> plan = new AtomicReference<>();
+    private final StreamingPlanSection sectionedPlan;
+    private final StageId rootStageId;
+    private final boolean summarizeTaskInfo;
+    private final int maxConcurrentMaterializations;
 
-    // The following fields are required for adaptive optimization in runtime.
+    // The following fields are required by adaptive optimization in runtime.
+    private final Session session;
     private final FunctionAndTypeManager functionAndTypeManager;
     private final List<PlanOptimizer> runtimePlanOptimizers;
     private final WarningCollector warningCollector;
     private final PlanNodeIdAllocator idAllocator;
     private final VariableAllocator variableAllocator;
+    private final SectionExecutionFactory sectionExecutionFactory;
+    private final RemoteTaskFactory remoteTaskFactory;
+    private final SplitSourceFactory splitSourceFactory;
     private final Set<StageId> runtimeOptimizedStages = Collections.synchronizedSet(new HashSet<>());
     private final PlanChecker planChecker;
     private final Metadata metadata;
     private final SqlParser sqlParser;
 
-    private final StreamingPlanSection sectionedPlan;
-    private final boolean summarizeTaskInfo;
-    private final int maxConcurrentMaterializations;
-    private final int maxStageRetries;
-
-    private final Map<StageId, List<SectionExecution>> sectionExecutions = new ConcurrentHashMap<>();
+    private final Map<StageId, StageExecutionAndScheduler> stageExecutions = new ConcurrentHashMap<>();
+    private final ExecutorService executor;
     private final AtomicBoolean started = new AtomicBoolean();
     private final AtomicBoolean scheduling = new AtomicBoolean();
-    private final AtomicInteger retriedSections = new AtomicInteger();
 
     private final PartialResultQueryTaskTracker partialResultQueryTaskTracker;
 
     public static SqlQueryScheduler createSqlQueryScheduler(
             LocationFactory locationFactory,
             ExecutionPolicy executionPolicy,
-            ExecutorService executor,
+            ExecutorService queryExecutor,
             SplitSchedulerStats schedulerStats,
             SectionExecutionFactory sectionExecutionFactory,
             RemoteTaskFactory remoteTaskFactory,
             SplitSourceFactory splitSourceFactory,
-            InternalNodeManager nodeManager,
             Session session,
+            FunctionAndTypeManager functionAndTypeManager,
             QueryStateMachine queryStateMachine,
             SubPlan plan,
+            OutputBuffers rootOutputBuffers,
             boolean summarizeTaskInfo,
-            FunctionAndTypeManager functionAndTypeManager,
             List<PlanOptimizer> runtimePlanOptimizers,
             WarningCollector warningCollector,
             PlanNodeIdAllocator idAllocator,
@@ -179,22 +167,22 @@ public class SqlQueryScheduler
             PlanChecker planChecker,
             Metadata metadata,
             SqlParser sqlParser,
-            PartialResultQueryManager partialResultQueriesHandler)
+            PartialResultQueryManager partialResultQueryManager)
     {
         SqlQueryScheduler sqlQueryScheduler = new SqlQueryScheduler(
                 locationFactory,
                 executionPolicy,
-                executor,
+                queryExecutor,
                 schedulerStats,
                 sectionExecutionFactory,
                 remoteTaskFactory,
                 splitSourceFactory,
-                nodeManager,
                 session,
+                functionAndTypeManager,
                 queryStateMachine,
                 plan,
                 summarizeTaskInfo,
-                functionAndTypeManager,
+                rootOutputBuffers,
                 runtimePlanOptimizers,
                 warningCollector,
                 idAllocator,
@@ -202,7 +190,7 @@ public class SqlQueryScheduler
                 planChecker,
                 metadata,
                 sqlParser,
-                partialResultQueriesHandler);
+                partialResultQueryManager);
         sqlQueryScheduler.initialize();
         return sqlQueryScheduler;
     }
@@ -210,17 +198,17 @@ public class SqlQueryScheduler
     private SqlQueryScheduler(
             LocationFactory locationFactory,
             ExecutionPolicy executionPolicy,
-            ExecutorService executor,
+            ExecutorService queryExecutor,
             SplitSchedulerStats schedulerStats,
             SectionExecutionFactory sectionExecutionFactory,
             RemoteTaskFactory remoteTaskFactory,
             SplitSourceFactory splitSourceFactory,
-            InternalNodeManager nodeManager,
             Session session,
+            FunctionAndTypeManager functionAndTypeManager,
             QueryStateMachine queryStateMachine,
             SubPlan plan,
             boolean summarizeTaskInfo,
-            FunctionAndTypeManager functionAndTypeManager,
+            OutputBuffers rootOutputBuffers,
             List<PlanOptimizer> runtimePlanOptimizers,
             WarningCollector warningCollector,
             PlanNodeIdAllocator idAllocator,
@@ -232,14 +220,11 @@ public class SqlQueryScheduler
     {
         this.locationFactory = requireNonNull(locationFactory, "locationFactory is null");
         this.executionPolicy = requireNonNull(executionPolicy, "schedulerPolicyFactory is null");
-        this.executor = requireNonNull(executor, "executor is null");
+        this.executor = queryExecutor;
         this.schedulerStats = requireNonNull(schedulerStats, "schedulerStats is null");
-        this.sectionExecutionFactory = requireNonNull(sectionExecutionFactory, "sectionExecutionFactory is null");
-        this.remoteTaskFactory = requireNonNull(remoteTaskFactory, "remoteTaskFactory is null");
-        this.splitSourceFactory = requireNonNull(splitSourceFactory, "splitSourceFactory is null");
-        this.nodeManager = requireNonNull(nodeManager, "nodeManager is null");
-        this.session = requireNonNull(session, "session is null");
         this.queryStateMachine = requireNonNull(queryStateMachine, "queryStateMachine is null");
+        this.plan.compareAndSet(null, requireNonNull(plan, "plan is null"));
+        this.session = requireNonNull(session, "session is null");
         this.functionAndTypeManager = requireNonNull(functionAndTypeManager, "functionManager is null");
         this.runtimePlanOptimizers = requireNonNull(runtimePlanOptimizers, "runtimePlanOptimizers is null");
         this.warningCollector = requireNonNull(warningCollector, "warningCollector is null");
@@ -248,17 +233,73 @@ public class SqlQueryScheduler
         this.planChecker = requireNonNull(planChecker, "planChecker is null");
         this.metadata = requireNonNull(metadata, "metadata is null");
         this.sqlParser = requireNonNull(sqlParser, "sqlParser is null");
-        this.plan.compareAndSet(null, requireNonNull(plan, "plan is null"));
+        this.sectionExecutionFactory = requireNonNull(sectionExecutionFactory, "sectionExecutionFactory is null");
+        this.remoteTaskFactory = requireNonNull(remoteTaskFactory, "remoteTaskFactory is null");
+        this.splitSourceFactory = requireNonNull(splitSourceFactory, "splitSourceFactory is null");
         this.sectionedPlan = extractStreamingSections(plan);
         this.summarizeTaskInfo = summarizeTaskInfo;
+
+        OutputBufferId rootBufferId = getOnlyElement(rootOutputBuffers.getBuffers().keySet());
+        List<StageExecutionAndScheduler> stageExecutions = createStageExecutions(
+                sectionExecutionFactory,
+                (fragmentId, tasks, noMoreExchangeLocations) -> updateQueryOutputLocations(queryStateMachine, rootBufferId, tasks, noMoreExchangeLocations),
+                sectionedPlan,
+                Optional.of(new int[1]),
+                rootOutputBuffers,
+                remoteTaskFactory,
+                splitSourceFactory,
+                session);
+
+        this.rootStageId = Iterables.getLast(stageExecutions).getStageExecution().getStageExecutionId().getStageId();
+
+        stageExecutions.stream()
+                .forEach(execution -> this.stageExecutions.put(execution.getStageExecution().getStageExecutionId().getStageId(), execution));
+
         this.maxConcurrentMaterializations = getMaxConcurrentMaterializations(session);
-        this.maxStageRetries = getMaxStageRetries(session);
         this.partialResultQueryTaskTracker = new PartialResultQueryTaskTracker(partialResultQueryManager, getPartialResultsCompletionRatioThreshold(session), getPartialResultsMaxExecutionTimeMultiplier(session), warningCollector);
     }
 
     // this is a separate method to ensure that the `this` reference is not leaked during construction
     private void initialize()
     {
+        SqlStageExecution rootStage = stageExecutions.get(rootStageId).getStageExecution();
+        rootStage.addStateChangeListener(state -> {
+            if (state == FINISHED) {
+                queryStateMachine.transitionToFinishing();
+            }
+            else if (state == CANCELED) {
+                // output stage was canceled
+                queryStateMachine.transitionToCanceled();
+            }
+        });
+
+        for (StageExecutionAndScheduler stageExecutionInfo : stageExecutions.values()) {
+            SqlStageExecution stageExecution = stageExecutionInfo.getStageExecution();
+            stageExecution.addStateChangeListener(state -> {
+                if (queryStateMachine.isDone()) {
+                    return;
+                }
+                if (state == FAILED) {
+                    queryStateMachine.transitionToFailed(stageExecution.getStageExecutionInfo().getFailureCause().get().toException());
+                }
+                else if (state == ABORTED) {
+                    // this should never happen, since abort can only be triggered in query clean up after the query is finished
+                    queryStateMachine.transitionToFailed(new PrestoException(GENERIC_INTERNAL_ERROR, "Query stage was aborted"));
+                }
+                else if (state == FINISHED) {
+                    // checks if there's any new sections available for execution and starts the scheduling if any
+                    startScheduling();
+                }
+                else if (queryStateMachine.getQueryState() == QueryState.STARTING) {
+                    // if the stage has at least one task, we are running
+                    if (stageExecution.hasTasks()) {
+                        queryStateMachine.transitionToRunning();
+                    }
+                }
+            });
+            stageExecution.addFinalStageInfoListener(status -> queryStateMachine.updateQueryInfo(Optional.of(getStageInfo())));
+        }
+
         // when query is done or any time a stage completes, attempt to transition query to "final query info ready"
         queryStateMachine.addStateChangeListener(newState -> {
             if (newState.isDone()) {
@@ -267,7 +308,64 @@ public class SqlQueryScheduler
         });
     }
 
-    @Override
+    private static void updateQueryOutputLocations(QueryStateMachine queryStateMachine, OutputBufferId rootBufferId, Set<RemoteTask> tasks, boolean noMoreExchangeLocations)
+    {
+        Map<URI, TaskId> bufferLocations = tasks.stream()
+                .collect(toImmutableMap(
+                        task -> getBufferLocation(task, rootBufferId),
+                        RemoteTask::getTaskId));
+        queryStateMachine.updateOutputLocations(bufferLocations, noMoreExchangeLocations);
+    }
+
+    private static URI getBufferLocation(RemoteTask remoteTask, OutputBufferId rootBufferId)
+    {
+        URI location = remoteTask.getTaskStatus().getSelf();
+        return uriBuilderFrom(location).appendPath("results").appendPath(rootBufferId.toString()).build();
+    }
+
+    /**
+     * returns a List of SqlStageExecutionInfos in a postorder representation of the tree
+     */
+    private List<StageExecutionAndScheduler> createStageExecutions(
+            SectionExecutionFactory sectionExecutionFactory,
+            ExchangeLocationsConsumer locationsConsumer,
+            StreamingPlanSection section,
+            Optional<int[]> bucketToPartition,
+            OutputBuffers outputBuffers,
+            RemoteTaskFactory remoteTaskFactory,
+            SplitSourceFactory splitSourceFactory,
+            Session session)
+    {
+        ImmutableList.Builder<StageExecutionAndScheduler> stages = ImmutableList.builder();
+
+        for (StreamingPlanSection childSection : section.getChildren()) {
+            ExchangeLocationsConsumer childLocationsConsumer = (fragmentId, tasks, noMoreExchangeLocations) -> {};
+            stages.addAll(createStageExecutions(
+                    sectionExecutionFactory,
+                    childLocationsConsumer,
+                    childSection,
+                    Optional.empty(),
+                    createDiscardingOutputBuffers(),
+                    remoteTaskFactory,
+                    splitSourceFactory,
+                    session));
+        }
+        List<StageExecutionAndScheduler> sectionStages =
+                sectionExecutionFactory.createSectionExecutions(
+                        session,
+                        section,
+                        locationsConsumer,
+                        bucketToPartition,
+                        outputBuffers,
+                        summarizeTaskInfo,
+                        remoteTaskFactory,
+                        splitSourceFactory,
+                        0).getSectionStages();
+        stages.addAll(sectionStages);
+
+        return stages.build();
+    }
+
     public void start()
     {
         if (started.compareAndSet(false, true)) {
@@ -277,6 +375,7 @@ public class SqlQueryScheduler
 
     private void startScheduling()
     {
+        requireNonNull(stageExecutions);
         // still scheduling the previous batch of stages
         if (scheduling.get()) {
             return;
@@ -296,51 +395,46 @@ public class SqlQueryScheduler
         try (SetThreadName ignored = new SetThreadName("Query-%s", queryStateMachine.getQueryId())) {
             Set<StageId> completedStages = new HashSet<>();
 
-            List<ExecutionSchedule> executionSchedules = new LinkedList<>();
+            List<ExecutionSchedule> sectionExecutionSchedules = new LinkedList<>();
 
             while (!Thread.currentThread().isInterrupted()) {
                 // remove finished section
-                executionSchedules.removeIf(ExecutionSchedule::isFinished);
+                sectionExecutionSchedules.removeIf(ExecutionSchedule::isFinished);
 
                 // try to pull more section that are ready to be run
                 List<StreamingPlanSection> sectionsReadyForExecution = getSectionsReadyForExecution();
 
                 // all finished
-                if (sectionsReadyForExecution.isEmpty() && executionSchedules.isEmpty()) {
+                if (sectionsReadyForExecution.isEmpty() && sectionExecutionSchedules.isEmpty()) {
                     break;
                 }
 
-                // Apply runtime CBO on the ready sections before creating SectionExecutions.
-                List<SectionExecution> sectionExecutions = createStageExecutions(sectionsReadyForExecution.stream()
-                        .map(this::tryCostBasedOptimize)
-                        .collect(toImmutableList()));
-                if (queryStateMachine.isDone()) {
-                    sectionExecutions.forEach(SectionExecution::abort);
-                    break;
-                }
-
-                sectionExecutions.forEach(sectionExecution -> scheduledStageExecutions.addAll(sectionExecution.getSectionStages()));
-                sectionExecutions.stream()
-                        .map(SectionExecution::getSectionStages)
+                List<List<StageExecutionAndScheduler>> sectionStageExecutions = getStageExecutions(sectionsReadyForExecution);
+                sectionStageExecutions.forEach(scheduledStageExecutions::addAll);
+                sectionStageExecutions.stream()
+                        .map(executionInfos -> executionInfos.stream()
+                                .collect(toImmutableList()))
                         .map(stages -> executionPolicy.createExecutionSchedule(session, stages))
-                        .forEach(executionSchedules::add);
+                        .forEach(sectionExecutionSchedules::add);
 
-                while (!executionSchedules.isEmpty() && executionSchedules.stream().noneMatch(ExecutionSchedule::isFinished)) {
+                while (sectionExecutionSchedules.stream().noneMatch(ExecutionSchedule::isFinished)) {
                     List<ListenableFuture<?>> blockedStages = new ArrayList<>();
 
-                    List<StageExecutionAndScheduler> executionsToSchedule = executionSchedules.stream()
+                    List<StageExecutionAndScheduler> executionsToSchedule = sectionExecutionSchedules.stream()
                             .flatMap(schedule -> schedule.getStagesToSchedule().stream())
                             .collect(toImmutableList());
 
-                    for (StageExecutionAndScheduler executionAndScheduler : executionsToSchedule) {
-                        executionAndScheduler.getStageExecution().beginScheduling();
+                    for (StageExecutionAndScheduler stageExecutionAndScheduler : executionsToSchedule) {
+                        SqlStageExecution stageExecution = stageExecutionAndScheduler.getStageExecution();
+                        StageId stageId = stageExecution.getStageExecutionId().getStageId();
+                        stageExecution.beginScheduling();
 
                         // perform some scheduling work
-                        ScheduleResult result = executionAndScheduler.getStageScheduler()
+                        ScheduleResult result = stageExecutionAndScheduler.getStageScheduler()
                                 .schedule();
 
                         // Track leaf tasks if partial results are enabled
-                        if (isPartialResultsEnabled(session) && executionAndScheduler.getStageExecution().getFragment().isLeaf()) {
+                        if (isPartialResultsEnabled(session) && stageExecutionAndScheduler.getStageExecution().getFragment().isLeaf()) {
                             for (RemoteTask task : result.getNewTasks()) {
                                 partialResultQueryTaskTracker.trackTask(task);
                                 task.addFinalTaskInfoListener(partialResultQueryTaskTracker::recordTaskFinish);
@@ -349,13 +443,13 @@ public class SqlQueryScheduler
 
                         // modify parent and children based on the results of the scheduling
                         if (result.isFinished()) {
-                            executionAndScheduler.getStageExecution().schedulingComplete();
+                            stageExecution.schedulingComplete();
                         }
                         else if (!result.getBlocked().isDone()) {
                             blockedStages.add(result.getBlocked());
                         }
-                        executionAndScheduler.getStageLinkage()
-                                .processScheduleResults(executionAndScheduler.getStageExecution().getState(), result.getNewTasks());
+                        stageExecutionAndScheduler.getStageLinkage()
+                                .processScheduleResults(stageExecution.getState(), result.getNewTasks());
                         schedulerStats.getSplitsScheduledPerIteration().add(result.getSplitsScheduled());
                         if (result.getBlockedReason().isPresent()) {
                             switch (result.getBlockedReason().get()) {
@@ -382,11 +476,11 @@ public class SqlQueryScheduler
 
                     // make sure to update stage linkage at least once per loop to catch async state changes (e.g., partial cancel)
                     boolean stageFinishedExecution = false;
-                    for (StageExecutionAndScheduler stageExecutionAndScheduler : scheduledStageExecutions) {
-                        SqlStageExecution stageExecution = stageExecutionAndScheduler.getStageExecution();
+                    for (StageExecutionAndScheduler stageExecutionInfo : scheduledStageExecutions) {
+                        SqlStageExecution stageExecution = stageExecutionInfo.getStageExecution();
                         StageId stageId = stageExecution.getStageExecutionId().getStageId();
                         if (!completedStages.contains(stageId) && stageExecution.getState().isDone()) {
-                            stageExecutionAndScheduler.getStageLinkage()
+                            stageExecutionInfo.getStageLinkage()
                                     .processScheduleResults(stageExecution.getState(), ImmutableSet.of());
                             completedStages.add(stageId);
                             stageFinishedExecution = true;
@@ -410,10 +504,10 @@ public class SqlQueryScheduler
                 }
             }
 
-            for (StageExecutionAndScheduler stageExecutionAndScheduler : scheduledStageExecutions) {
-                StageExecutionState state = stageExecutionAndScheduler.getStageExecution().getState();
+            for (StageExecutionAndScheduler stageExecutionInfo : scheduledStageExecutions) {
+                StageExecutionState state = stageExecutionInfo.getStageExecution().getState();
                 if (state != SCHEDULED && state != RUNNING && !state.isDone()) {
-                    throw new PrestoException(GENERIC_INTERNAL_ERROR, format("Scheduling is complete, but stage execution %s is in state %s", stageExecutionAndScheduler.getStageExecution().getStageExecutionId(), state));
+                    throw new PrestoException(GENERIC_INTERNAL_ERROR, format("Scheduling is complete, but stage execution %s is in state %s", stageExecutionInfo.getStageExecution().getStageExecutionId(), state));
                 }
             }
 
@@ -433,9 +527,9 @@ public class SqlQueryScheduler
         }
         finally {
             RuntimeException closeError = new RuntimeException();
-            for (StageExecutionAndScheduler stageExecutionAndScheduler : scheduledStageExecutions) {
+            for (StageExecutionAndScheduler stageExecutionInfo : scheduledStageExecutions) {
                 try {
-                    stageExecutionAndScheduler.getStageScheduler().close();
+                    stageExecutionInfo.getStageScheduler().close();
                 }
                 catch (Throwable t) {
                     queryStateMachine.transitionToFailed(t);
@@ -451,6 +545,21 @@ public class SqlQueryScheduler
         }
     }
 
+    private List<StreamingPlanSection> getSectionsReadyForExecution()
+    {
+        long runningPlanSections =
+                stream(forTree(StreamingPlanSection::getChildren).depthFirstPreOrder(sectionedPlan))
+                        .map(section -> getStageExecution(section.getPlan().getFragment().getId()).getState())
+                        .filter(state -> !state.isDone() && state != PLANNED)
+                        .count();
+        return stream(forTree(StreamingPlanSection::getChildren).depthFirstPreOrder(sectionedPlan))
+                // get all sections ready for execution
+                .filter(this::isReadyForExecution)
+                .limit(maxConcurrentMaterializations - runningPlanSections)
+                .map(this::tryCostBasedOptimize)
+                .collect(toImmutableList());
+    }
+
     /**
      * A general purpose utility function to invoke runtime cost-based optimizer.
      * (right now there is only one plan optimizer which determines if the probe and build side of a JoinNode should be swapped
@@ -463,7 +572,7 @@ public class SqlQueryScheduler
             return section;
         }
 
-        // Apply runtime optimization on each StreamingSubPlan's fragment
+        // Apply runtime optimization on each StreamingSubPlan and generate optimized new fragments
         Map<PlanFragment, PlanFragment> oldToNewFragment = new HashMap<>();
         stream(forTree(StreamingSubPlan::getChildren).depthFirstPreOrder(section.getPlan()))
                 .forEach(currentSubPlan -> {
@@ -484,11 +593,13 @@ public class SqlQueryScheduler
         // Update SubPlan so that getStageInfo will reflect the latest optimized plan when query is finished.
         updatePlan(oldToNewFragment);
 
+        // Rebuild and update entries of the stageExecutions map.
+        updateStageExecutions(section, oldToNewFragment);
         log.debug("Invoked CBO during runtime, optimized stage IDs: " + oldToNewFragment.keySet().stream()
                 .map(PlanFragment::getId)
                 .map(PlanFragmentId::toString)
                 .collect(Collectors.joining(", ")));
-        return new StreamingPlanSection(rewriteStreamingSubPlan(section.getPlan(), oldToNewFragment), section.getChildren());
+        return section;
     }
 
     private Optional<PlanFragment> performRuntimeOptimizations(StreamingSubPlan subPlan)
@@ -518,6 +629,48 @@ public class SqlQueryScheduler
         return Optional.empty();
     }
 
+    /**
+     * Utility function that rebuild a StreamingPlanSection, re-create stageExecutionAndScheduler for each of its stage, and finally update the stageExecutions map.
+     */
+    private void updateStageExecutions(StreamingPlanSection section, Map<PlanFragment, PlanFragment> oldToNewFragment)
+    {
+        StreamingPlanSection newSection = new StreamingPlanSection(rewriteStreamingSubPlan(section.getPlan(), oldToNewFragment), section.getChildren());
+        PlanFragment sectionRootFragment = newSection.getPlan().getFragment();
+        Optional<int[]> bucketToPartition;
+        OutputBuffers outputBuffers;
+        ExchangeLocationsConsumer locationsConsumer;
+        if (isRootFragment(sectionRootFragment)) {
+            bucketToPartition = Optional.of(new int[1]);
+            outputBuffers = createInitialEmptyOutputBuffers(sectionRootFragment.getPartitioningScheme().getPartitioning().getHandle())
+                    .withBuffer(new OutputBufferId(0), BROADCAST_PARTITION_ID)
+                    .withNoMoreBufferIds();
+            OutputBufferId rootBufferId = getOnlyElement(outputBuffers.getBuffers().keySet());
+            locationsConsumer = (fragmentId, tasks, noMoreExchangeLocations) ->
+                    updateQueryOutputLocations(queryStateMachine, rootBufferId, tasks, noMoreExchangeLocations);
+        }
+        else {
+            bucketToPartition = Optional.empty();
+            outputBuffers = createDiscardingOutputBuffers();
+            locationsConsumer = (fragmentId, tasks, noMoreExchangeLocations) -> {};
+        }
+        SectionExecution sectionExecution = sectionExecutionFactory.createSectionExecutions(
+                session,
+                newSection,
+                locationsConsumer,
+                bucketToPartition,
+                outputBuffers,
+                summarizeTaskInfo,
+                remoteTaskFactory,
+                splitSourceFactory,
+                0);
+        addStateChangeListeners(sectionExecution);
+        Map<StageId, StageExecutionAndScheduler> updatedStageExecutions = sectionExecution.getSectionStages().stream()
+                .collect(toImmutableMap(execution -> execution.getStageExecution().getStageExecutionId().getStageId(), identity()));
+        synchronized (this) {
+            stageExecutions.putAll(updatedStageExecutions);
+        }
+    }
+
     private void updatePlan(Map<PlanFragment, PlanFragment> oldToNewFragments)
     {
         plan.getAndUpdate(value -> rewritePlan(value, oldToNewFragments));
@@ -537,131 +690,7 @@ public class SqlQueryScheduler
         }
     }
 
-    private StreamingSubPlan rewriteStreamingSubPlan(StreamingSubPlan root, Map<PlanFragment, PlanFragment> oldToNewFragment)
-    {
-        ImmutableList.Builder<StreamingSubPlan> childrenPlans = ImmutableList.builder();
-        for (StreamingSubPlan child : root.getChildren()) {
-            childrenPlans.add(rewriteStreamingSubPlan(child, oldToNewFragment));
-        }
-        if (oldToNewFragment.containsKey(root.getFragment())) {
-            return new StreamingSubPlan(oldToNewFragment.get(root.getFragment()), childrenPlans.build());
-        }
-        else {
-            return new StreamingSubPlan(root.getFragment(), childrenPlans.build());
-        }
-    }
-
-    private List<StreamingPlanSection> getSectionsReadyForExecution()
-    {
-        long runningPlanSections =
-                stream(forTree(StreamingPlanSection::getChildren).depthFirstPreOrder(sectionedPlan))
-                        .map(section -> getLatestSectionExecution(getStageId(section.getPlan().getFragment().getId())))
-                        .filter(Optional::isPresent)
-                        .map(Optional::get)
-                        .filter(SectionExecution::isRunning)
-                        .count();
-        return stream(forTree(StreamingPlanSection::getChildren).depthFirstPreOrder(sectionedPlan))
-                // get all sections ready for execution
-                .filter(this::isReadyForExecution)
-                .limit(maxConcurrentMaterializations - runningPlanSections)
-                .collect(toImmutableList());
-    }
-
-    private boolean isReadyForExecution(StreamingPlanSection section)
-    {
-        Optional<SectionExecution> sectionExecution = getLatestSectionExecution(getStageId(section.getPlan().getFragment().getId()));
-        if (sectionExecution.isPresent() && (sectionExecution.get().isRunning() || sectionExecution.get().isFinished())) {
-            // already scheduled
-            return false;
-        }
-        for (StreamingPlanSection child : section.getChildren()) {
-            Optional<SectionExecution> childSectionExecution = getLatestSectionExecution(getStageId(child.getPlan().getFragment().getId()));
-            if (!childSectionExecution.isPresent() || !childSectionExecution.get().isFinished()) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    private Optional<SectionExecution> getLatestSectionExecution(StageId stageId)
-    {
-        List<SectionExecution> sectionExecutions = this.sectionExecutions.get(stageId);
-        if (sectionExecutions == null || sectionExecutions.isEmpty()) {
-            return Optional.empty();
-        }
-        return Optional.of(getLast(sectionExecutions));
-    }
-
-    private StageId getStageId(PlanFragmentId fragmentId)
-    {
-        return new StageId(session.getQueryId(), fragmentId.getId());
-    }
-
-    private List<SectionExecution> createStageExecutions(List<StreamingPlanSection> sections)
-    {
-        ImmutableList.Builder<SectionExecution> result = ImmutableList.builder();
-        for (StreamingPlanSection section : sections) {
-            StageId sectionId = getStageId(section.getPlan().getFragment().getId());
-            List<SectionExecution> attempts = sectionExecutions.computeIfAbsent(sectionId, (ignored) -> new CopyOnWriteArrayList<>());
-
-            // sectionExecutions only get created when they are about to be scheduled, so there should
-            // never be a non-failed SectionExecution for a section that's ready for execution
-            verify(attempts.isEmpty() || getLast(attempts).isFailed(), "Non-failed sectionExecutions already exists");
-
-            PlanFragment sectionRootFragment = section.getPlan().getFragment();
-
-            Optional<int[]> bucketToPartition;
-            OutputBuffers outputBuffers;
-            ExchangeLocationsConsumer locationsConsumer;
-            if (isRootFragment(sectionRootFragment)) {
-                bucketToPartition = Optional.of(new int[1]);
-                outputBuffers = createInitialEmptyOutputBuffers(sectionRootFragment.getPartitioningScheme().getPartitioning().getHandle())
-                        .withBuffer(new OutputBufferId(0), BROADCAST_PARTITION_ID)
-                        .withNoMoreBufferIds();
-                OutputBufferId rootBufferId = getOnlyElement(outputBuffers.getBuffers().keySet());
-                locationsConsumer = (fragmentId, tasks, noMoreExchangeLocations) ->
-                        updateQueryOutputLocations(queryStateMachine, rootBufferId, tasks, noMoreExchangeLocations);
-            }
-            else {
-                bucketToPartition = Optional.empty();
-                outputBuffers = createDiscardingOutputBuffers();
-                locationsConsumer = (fragmentId, tasks, noMoreExchangeLocations) -> {};
-            }
-
-            int attemptId = attempts.size();
-            SectionExecution sectionExecution = sectionExecutionFactory.createSectionExecutions(
-                    session,
-                    section,
-                    locationsConsumer,
-                    bucketToPartition,
-                    outputBuffers,
-                    summarizeTaskInfo,
-                    remoteTaskFactory,
-                    splitSourceFactory,
-                    attemptId);
-
-            addStateChangeListeners(sectionExecution);
-            attempts.add(sectionExecution);
-            result.add(sectionExecution);
-        }
-        return result.build();
-    }
-
-    private static void updateQueryOutputLocations(QueryStateMachine queryStateMachine, OutputBufferId rootBufferId, Set<RemoteTask> tasks, boolean noMoreExchangeLocations)
-    {
-        Map<URI, TaskId> bufferLocations = tasks.stream()
-                .collect(toImmutableMap(
-                        task -> getBufferLocation(task, rootBufferId),
-                        RemoteTask::getTaskId));
-        queryStateMachine.updateOutputLocations(bufferLocations, noMoreExchangeLocations);
-    }
-
-    private static URI getBufferLocation(RemoteTask remoteTask, OutputBufferId rootBufferId)
-    {
-        URI location = remoteTask.getTaskStatus().getSelf();
-        return uriBuilderFrom(location).appendPath("results").appendPath(rootBufferId.toString()).build();
-    }
-
+    // Only used for adaptive optimization, to register listeners to new stageExecutions generated in runtime.
     private void addStateChangeListeners(SectionExecution sectionExecution)
     {
         for (StageExecutionAndScheduler stageExecutionAndScheduler : sectionExecution.getSectionStages()) {
@@ -681,42 +710,12 @@ public class SqlQueryScheduler
                 if (queryStateMachine.isDone()) {
                     return;
                 }
-
                 if (state == FAILED) {
-                    ExecutionFailureInfo failureInfo = stageExecution.getStageExecutionInfo().getFailureCause()
-                            .orElseThrow(() -> new VerifyException(format("stage execution failed, but the failure info is missing: %s", stageExecution.getStageExecutionId())));
-                    Exception failureException = failureInfo.toException();
-
-                    boolean isRootSection = isRootFragment(sectionExecution.getRootStage().getStageExecution().getFragment());
-                    // root section directly streams the results to the user, cannot be retried
-                    if (isRootSection) {
-                        queryStateMachine.transitionToFailed(failureException);
-                        return;
-                    }
-
-                    if (retriedSections.get() >= maxStageRetries) {
-                        queryStateMachine.transitionToFailed(failureException);
-                        return;
-                    }
-
-                    if (!RECOVERABLE_ERROR_CODES.contains(failureInfo.getErrorCode())) {
-                        queryStateMachine.transitionToFailed(failureException);
-                        return;
-                    }
-
-                    try {
-                        if (sectionExecution.abort()) {
-                            retriedSections.incrementAndGet();
-                            nodeManager.refreshNodes();
-                            startScheduling();
-                        }
-                    }
-                    catch (Throwable t) {
-                        if (failureException != t) {
-                            failureException.addSuppressed(t);
-                        }
-                        queryStateMachine.transitionToFailed(failureException);
-                    }
+                    queryStateMachine.transitionToFailed(stageExecution.getStageExecutionInfo().getFailureCause().get().toException());
+                }
+                else if (state == ABORTED) {
+                    // this should never happen, since abort can only be triggered in query clean up after the query is finished
+                    queryStateMachine.transitionToFailed(new PrestoException(GENERIC_INTERNAL_ERROR, "Query stage was aborted"));
                 }
                 else if (state == FINISHED) {
                     // checks if there's any new sections available for execution and starts the scheduling if any
@@ -733,24 +732,86 @@ public class SqlQueryScheduler
         }
     }
 
-    @Override
+    private StreamingSubPlan rewriteStreamingSubPlan(StreamingSubPlan root, Map<PlanFragment, PlanFragment> oldToNewFragment)
+    {
+        ImmutableList.Builder<StreamingSubPlan> childrenPlans = ImmutableList.builder();
+        for (StreamingSubPlan child : root.getChildren()) {
+            childrenPlans.add(rewriteStreamingSubPlan(child, oldToNewFragment));
+        }
+        if (oldToNewFragment.containsKey(root.getFragment())) {
+            return new StreamingSubPlan(oldToNewFragment.get(root.getFragment()), childrenPlans.build());
+        }
+        else {
+            return new StreamingSubPlan(root.getFragment(), childrenPlans.build());
+        }
+    }
+
+    private static boolean isRootFragment(PlanFragment fragment)
+    {
+        return fragment.getId().getId() == ROOT_FRAGMENT_ID;
+    }
+
+    private boolean isReadyForExecution(StreamingPlanSection section)
+    {
+        SqlStageExecution stageExecution = getStageExecution(section.getPlan().getFragment().getId());
+        if (stageExecution.getState() != PLANNED) {
+            // already scheduled
+            return false;
+        }
+        for (StreamingPlanSection child : section.getChildren()) {
+            SqlStageExecution rootStageExecution = getStageExecution(child.getPlan().getFragment().getId());
+            if (rootStageExecution.getState() != FINISHED) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    private List<List<StageExecutionAndScheduler>> getStageExecutions(List<StreamingPlanSection> sections)
+    {
+        return sections.stream()
+                .map(section -> stream(forTree(StreamingSubPlan::getChildren).depthFirstPreOrder(section.getPlan())).collect(toImmutableList()))
+                .map(plans -> plans.stream()
+                        .map(StreamingSubPlan::getFragment)
+                        .map(PlanFragment::getId)
+                        .map(this::getStageExecutionInfo)
+                        .collect(toImmutableList()))
+                .collect(toImmutableList());
+    }
+
+    private SqlStageExecution getStageExecution(PlanFragmentId planFragmentId)
+    {
+        return stageExecutions.get(getStageId(planFragmentId)).getStageExecution();
+    }
+
+    private StageExecutionAndScheduler getStageExecutionInfo(PlanFragmentId planFragmentId)
+    {
+        return stageExecutions.get(getStageId(planFragmentId));
+    }
+
+    private StageId getStageId(PlanFragmentId fragmentId)
+    {
+        return new StageId(queryStateMachine.getQueryId(), fragmentId.getId());
+    }
+
     public long getUserMemoryReservation()
     {
-        return getAllStagesExecutions().mapToLong(SqlStageExecution::getUserMemoryReservation).sum();
+        return stageExecutions.values().stream()
+                .mapToLong(stageExecutionInfo -> stageExecutionInfo.getStageExecution().getUserMemoryReservation())
+                .sum();
     }
 
-    @Override
     public long getTotalMemoryReservation()
     {
-        return getAllStagesExecutions().mapToLong(SqlStageExecution::getTotalMemoryReservation).sum();
+        return stageExecutions.values().stream()
+                .mapToLong(stageExecutionInfo -> stageExecutionInfo.getStageExecution().getTotalMemoryReservation())
+                .sum();
     }
 
-    @Override
     public Duration getTotalCpuTime()
     {
-        long millis = getAllStagesExecutions()
-                .map(SqlStageExecution::getTotalCpuTime)
-                .mapToLong(Duration::toMillis)
+        long millis = stageExecutions.values().stream()
+                .mapToLong(stage -> stage.getStageExecution().getTotalCpuTime().toMillis())
                 .sum();
         return new Duration(millis, MILLISECONDS);
     }
@@ -758,104 +819,72 @@ public class SqlQueryScheduler
     @Override
     public DataSize getRawInputDataSize()
     {
-        long rawInputDataSize = getAllStagesExecutions()
-                .map(SqlStageExecution::getRawInputDataSize)
-                .mapToLong(DataSize::toBytes)
+        long datasize = stageExecutions.values().stream()
+                .mapToLong(stage -> stage.getStageExecution().getRawInputDataSize().toBytes())
                 .sum();
-        return DataSize.succinctBytes(rawInputDataSize);
+        return DataSize.succinctBytes(datasize);
     }
 
     @Override
     public long getOutputPositions()
     {
-        return getStageInfo().getLatestAttemptExecutionInfo().getStats().getOutputPositions();
+        return stageExecutions.get(rootStageId).getStageExecution().getStageExecutionInfo().getStats().getOutputPositions();
     }
 
     @Override
     public DataSize getOutputDataSize()
     {
-        return getStageInfo().getLatestAttemptExecutionInfo().getStats().getOutputDataSize();
+        return stageExecutions.get(rootStageId).getStageExecution().getStageExecutionInfo().getStats().getOutputDataSize();
     }
 
-    @Override
     public BasicStageExecutionStats getBasicStageStats()
     {
-        List<BasicStageExecutionStats> stageStats = getAllStagesExecutions()
-                .map(SqlStageExecution::getBasicStageStats)
+        List<BasicStageExecutionStats> stageStats = stageExecutions.values().stream()
+                .map(stageExecutionInfo -> stageExecutionInfo.getStageExecution().getBasicStageStats())
                 .collect(toImmutableList());
+
         return aggregateBasicStageStats(stageStats);
     }
 
-    private Stream<SqlStageExecution> getAllStagesExecutions()
-    {
-        return sectionExecutions.values().stream()
-                .flatMap(Collection::stream)
-                .flatMap(sectionExecution -> sectionExecution.getSectionStages().stream())
-                .map(StageExecutionAndScheduler::getStageExecution);
-    }
-
-    @Override
     public StageInfo getStageInfo()
     {
-        ListMultimap<StageId, SqlStageExecution> stageExecutions = getStageExecutions();
-        return buildStageInfo(plan.get(), stageExecutions);
+        Map<StageId, StageExecutionInfo> stageInfos = stageExecutions.values().stream()
+                .map(StageExecutionAndScheduler::getStageExecution)
+                .collect(toImmutableMap(execution -> execution.getStageExecutionId().getStageId(), SqlStageExecution::getStageExecutionInfo));
+
+        return buildStageInfo(plan.get(), stageInfos);
     }
 
-    private StageInfo buildStageInfo(SubPlan subPlan, ListMultimap<StageId, SqlStageExecution> stageExecutions)
+    private StageInfo buildStageInfo(SubPlan subPlan, Map<StageId, StageExecutionInfo> stageExecutionInfos)
     {
         StageId stageId = getStageId(subPlan.getFragment().getId());
-        List<SqlStageExecution> attempts = stageExecutions.get(stageId);
-
-        StageExecutionInfo latestAttemptInfo = attempts.isEmpty() ?
-                unscheduledExecutionInfo(stageId.getId(), queryStateMachine.isDone()) :
-                getLast(attempts).getStageExecutionInfo();
-        List<StageExecutionInfo> previousAttemptInfos = attempts.size() < 2 ?
-                ImmutableList.of() :
-                attempts.subList(0, attempts.size() - 1).stream()
-                        .map(SqlStageExecution::getStageExecutionInfo)
-                        .collect(toImmutableList());
-
+        StageExecutionInfo stageExecutionInfo = stageExecutionInfos.get(stageId);
+        checkArgument(stageExecutionInfo != null, "No stageExecutionInfo for %s", stageId);
         return new StageInfo(
                 stageId,
                 locationFactory.createStageLocation(stageId),
                 Optional.of(subPlan.getFragment()),
-                latestAttemptInfo,
-                previousAttemptInfos,
+                stageExecutionInfo,
+                ImmutableList.of(),
                 subPlan.getChildren().stream()
-                        .map(plan -> buildStageInfo(plan, stageExecutions))
+                        .map(plan -> buildStageInfo(plan, stageExecutionInfos))
                         .collect(toImmutableList()),
                 runtimeOptimizedStages.contains(stageId));
     }
 
-    private ListMultimap<StageId, SqlStageExecution> getStageExecutions()
-    {
-        ImmutableListMultimap.Builder<StageId, SqlStageExecution> result = ImmutableListMultimap.builder();
-        for (Collection<SectionExecution> sectionExecutionAttempts : sectionExecutions.values()) {
-            for (SectionExecution sectionExecution : sectionExecutionAttempts) {
-                for (StageExecutionAndScheduler stageExecution : sectionExecution.getSectionStages()) {
-                    result.put(stageExecution.getStageExecution().getStageExecutionId().getStageId(), stageExecution.getStageExecution());
-                }
-            }
-        }
-        return result.build();
-    }
-
-    @Override
     public void cancelStage(StageId stageId)
     {
         try (SetThreadName ignored = new SetThreadName("Query-%s", queryStateMachine.getQueryId())) {
-            getAllStagesExecutions()
-                    .filter(execution -> execution.getStageExecutionId().getStageId().equals(stageId))
-                    .forEach(SqlStageExecution::cancel);
+            SqlStageExecution execution = stageExecutions.get(stageId).getStageExecution();
+            SqlStageExecution stage = requireNonNull(execution, () -> format("Stage %s does not exist", stageId));
+            stage.cancel();
         }
     }
 
-    @Override
     public void abort()
     {
         try (SetThreadName ignored = new SetThreadName("Query-%s", queryStateMachine.getQueryId())) {
-            checkState(queryStateMachine.isDone(), "query scheduler is expected to be aborted only if the query is finished: %s", queryStateMachine.getQueryState());
-            getAllStagesExecutions().forEach(SqlStageExecution::abort);
+            stageExecutions.values().forEach(stageExecutionInfo -> stageExecutionInfo.getStageExecution().abort());
         }
     }
 }
