+++ presto-main/src/main/java/com/facebook/presto/sql/analyzer/FeaturesConfig.java
@@ -57,7 +57,9 @@ import static java.util.concurrent.TimeUnit.SECONDS;
         "analyzer.experimental-syntax-enabled",
         "optimizer.processing-optimization",
         "deprecated.legacy-order-by",
-        "deprecated.legacy-join-using"})
+        "deprecated.legacy-join-using",
+        "use-legacy-scheduler",
+        "max-stage-retries"})
 public class FeaturesConfig
 {
     @VisibleForTesting
@@ -79,7 +81,6 @@ public class FeaturesConfig
     private boolean groupedExecutionEnabled = true;
     private boolean recoverableGroupedExecutionEnabled;
     private double maxFailedTaskPercentage = 0.3;
-    private int maxStageRetries;
     private int concurrentLifespansPerTask;
     private boolean spatialJoinsEnabled = true;
     private boolean fastInequalityJoins = true;
@@ -207,7 +208,6 @@ public class FeaturesConfig
 
     private boolean listBuiltInFunctionsOnly = true;
     private boolean experimentalFunctionsEnabled;
-    private boolean useLegacyScheduler = true;
     private boolean optimizeCommonSubExpressions = true;
     private boolean preferDistributedUnion = true;
     private boolean optimizeNullsInJoin;
@@ -715,19 +715,6 @@ public class FeaturesConfig
     public FeaturesConfig setMaxFailedTaskPercentage(double maxFailedTaskPercentage)
     {
         this.maxFailedTaskPercentage = maxFailedTaskPercentage;
-        return this;
-    }
-
-    public int getMaxStageRetries()
-    {
-        return maxStageRetries;
-    }
-
-    @Config("max-stage-retries")
-    @ConfigDescription("Maximum number of times that stages can be retried")
-    public FeaturesConfig setMaxStageRetries(int maxStageRetries)
-    {
-        this.maxStageRetries = maxStageRetries;
         return this;
     }
 
@@ -2036,19 +2023,6 @@ public class FeaturesConfig
     public FeaturesConfig setExperimentalFunctionsEnabled(boolean experimentalFunctionsEnabled)
     {
         this.experimentalFunctionsEnabled = experimentalFunctionsEnabled;
-        return this;
-    }
-
-    public boolean isUseLegacyScheduler()
-    {
-        return useLegacyScheduler;
-    }
-
-    @Config("use-legacy-scheduler")
-    @ConfigDescription("Use the version of the scheduler before refactorings for section retries")
-    public FeaturesConfig setUseLegacyScheduler(boolean useLegacyScheduler)
-    {
-        this.useLegacyScheduler = useLegacyScheduler;
         return this;
     }
